+++ Plugins/UVOCBot.Plugins.Planetside/Abstractions/Services/ICensusApiService.cs
@@ -105,12 +105,36 @@ public interface ICensusApiService
     /// <summary>
     /// Gets a collection of outfit war registrations for the given world.
     /// </summary>
-    /// <param name="world">The world.</param>
+    /// <param name="outfitWarID">The ID of the war to retrieve registrations for.</param>
     /// <param name="ct">A <see cref="CancellationToken"/> that can be used to stop the operation.</param>
     /// <returns>A result representing the outcome of the operation, and containing the outfit registrations if successful.</returns>
     Task<Result<List<OutfitWarRegistration>>> GetOutfitWarRegistrationsAsync
     (
+        uint outfitWarID,
+        CancellationToken ct = default
+    );
+
+    /// <summary>
+    /// Gets information about the current outfit war for the given world.
+    /// </summary>
+    /// <param name="world">The world.</param>
+    /// <param name="ct">A <see cref="CancellationToken"/> that can be used to stop the operation.</param>
+    /// <returns>The current <see cref="OutfitWar"/>, or <c>null</c> if there is none.</returns>
+    Task<Result<OutfitWar?>> GetCurrentOutfitWar
+    (
         ValidWorldDefinition world,
+        CancellationToken ct = default
+    );
+
+    /// <summary>
+    /// Gets the matches of the current round of the given outfit war.
+    /// </summary>
+    /// <param name="outfitWarID">The ID of the war to retrieve the current matches of.</param>
+    /// <param name="ct">A <see cref="CancellationToken"/> that can be used to stop the operation.</param>
+    /// <returns>The current matches, or <c>null</c> if the war does not have a round in progress.</returns>
+    Task<Result<OutfitWarRoundWithMatches?>> GetCurrentOutfitWarMatches
+    (
+        uint outfitWarID,
         CancellationToken ct = default
     );
 }