+++ Plugins/UVOCBot.Plugins.Planetside/Objects/SanctuaryCensus/OutfitWar.cs
@@ -1,0 +1,15 @@
+using DbgCensus.Core.Objects;
+using System;
+
+namespace UVOCBot.Plugins.Planetside.Objects.SanctuaryCensus;
+
+public record OutfitWar
+(
+    uint OutfitWarID,
+    WorldDefinition WorldID,
+    GlobalizedString Title,
+    uint OutfitSignupRequirement,
+    DateTimeOffset StartTime,
+    DateTimeOffset EndTime,
+    string ImagePath
+);