+++ Plugins/UVOCBot.Plugins.Planetside/Services/CensusApiService.cs
@@ -245,15 +245,51 @@ public class CensusApiService : ICensusApiService, IDisposable
     /// <inheritdoc />
     public virtual async Task<Result<List<OutfitWarRegistration>>> GetOutfitWarRegistrationsAsync
     (
-        ValidWorldDefinition world,
+        uint outfitWarID,
         CancellationToken ct = default
     )
     {
         IQueryBuilder query = _queryService.CreateQuery(_sanctuaryOptions)
             .OnCollection("outfit_war_registration")
-            .Where("world_id", SearchModifier.Equals, (uint)world);
+            .Where("outfit_war_id", SearchModifier.Equals, outfitWarID);
 
         return await GetListAsync<OutfitWarRegistration>(query, ct).ConfigureAwait(false);
+    }
+
+    /// <inheritdoc />
+    public virtual async Task<Result<OutfitWar?>> GetCurrentOutfitWar
+    (
+        ValidWorldDefinition world,
+        CancellationToken ct = default
+    )
+    {
+        IQueryBuilder query = _queryService.CreateQuery(_sanctuaryOptions)
+            .OnCollection("outfit_war")
+            .Where("world_id", SearchModifier.Equals, (uint)world)
+            .Where("end_time", SearchModifier.GreaterThanOrEqual, DateTimeOffset.UtcNow.ToUnixTimeSeconds());
+
+        return await GetAsync<OutfitWar>(query, ct).ConfigureAwait(false);
+    }
+
+    /// <inheritdoc />
+    public virtual async Task<Result<OutfitWarRoundWithMatches?>> GetCurrentOutfitWarMatches
+    (
+        uint outfitWarID,
+        CancellationToken ct = default
+    )
+    {
+        IQueryBuilder query = _queryService.CreateQuery(_sanctuaryOptions)
+            .OnCollection("outfit_war_round")
+            .Where("outfit_war_id", SearchModifier.Equals, outfitWarID)
+            .Where("start_time", SearchModifier.LessThanOrEqual, DateTimeOffset.UtcNow.ToUnixTimeSeconds())
+            .Where("end_time", SearchModifier.GreaterThanOrEqual, DateTimeOffset.UtcNow.ToUnixTimeSeconds())
+            .AddJoin("outfit_war_match", j =>
+            {
+                j.IsList()
+                    .InjectAt("matches");
+            });
+
+        return await GetAsync<OutfitWarRoundWithMatches>(query, ct).ConfigureAwait(false);
     }
 
     /// <inheritdoc />